requirements.txt
discord.py
aioconsole
./lkk_log_pack